+++ example/example_rest/bin/person.g.dart
@@ -1,0 +1,23 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'person.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+Person _$PersonFromJson(Map<String, dynamic> json) {
+  return Person(
+    id: json['id'] as String?,
+    firstName: json['firstName'] as String,
+    lastName: json['lastName'] as String,
+    age: json['age'] as int,
+  );
+}
+
+Map<String, dynamic> _$PersonToJson(Person instance) => <String, dynamic>{
+      'id': instance.id,
+      'firstName': instance.firstName,
+      'lastName': instance.lastName,
+      'age': instance.age,
+    };