+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
-- **New:** Request Body Accessor
+- Added mechanism for request body handling
+- JSON handler can now process `Iterable`s and serialize model classes of `Iterable`s
 
 ## 0.0.1